Reorganize wasm crate
