+++ crates/hyperdrive-wasm/src/long.rs
@@ -1,0 +1,66 @@
+use crate::utils::set_panic_hook;
+use crate::{JsPoolConfig, JsPoolInfo};
+use ethers::types::{I256, U256};
+use fixed_point::FixedPoint;
+use hyperdrive_math::State;
+use wasm_bindgen::prelude::*;
+
+/// Gets the long amount that will be opened for a given base amount.
+///
+/// @param poolInfo - The current state of the pool
+///
+/// @param poolConfig - The pool's configuration
+///
+/// @param baseAmount - The amount of base tokens to open a long for
+#[wasm_bindgen(skip_jsdoc)]
+pub fn calcOpenLong(
+    poolInfo: &JsPoolInfo,
+    poolConfig: &JsPoolConfig,
+    baseAmount: String,
+) -> String {
+    set_panic_hook();
+    let state = State {
+        info: poolInfo.into(),
+        config: poolConfig.into(),
+    };
+    let base_amount: FixedPoint = FixedPoint::from(U256::from_dec_str(&baseAmount).unwrap());
+
+    state.calculate_open_long(base_amount).to_string()
+}
+
+/// Gets the max long that can be opened given a budget.
+///
+/// @param poolInfo - The current state of the pool
+///
+/// @param poolConfig - The pool's configuration
+///
+/// @param budget - The maximum amount of base tokens that can be spent
+///
+/// @param checkpointExposure - The exposure of the pool's current checkpoint
+///
+/// @param maybeMaxIterations - The maximum number of iterations to run the
+/// binary search for
+#[wasm_bindgen(skip_jsdoc)]
+pub fn getMaxLong(
+    poolInfo: &JsPoolInfo,
+    poolConfig: &JsPoolConfig,
+    budget: String,
+    checkpointExposure: String,
+    maybeMaxIterations: Option<u8>,
+) -> String {
+    set_panic_hook();
+    let state = State {
+        info: poolInfo.into(),
+        config: poolConfig.into(),
+    };
+    let _budget = U256::from_dec_str(&budget).unwrap();
+    let checkpoint_exposure: I256 = I256::from_dec_str(&checkpointExposure).unwrap();
+
+    state
+        .get_max_long(
+            _budget,
+            checkpoint_exposure,
+            maybeMaxIterations.map(|x| x.into()),
+        )
+        .to_string()
+}
